monitor/testcard: Make it portable
0xbf44101c and 0xbf441020 are part of the display controller's MMIO
space and contain the chroma and luma buffer base addresses, albeit
shifted by three bits (which made them unnecessarily difficult to find).

Hardcoding the buffer addresses meant that we depended on the exact
behavior of the memory allocator in the "image" program as well as the
boot splash configured. Now, testcard.c should be portable to all
M88CS8001-based devices where the framebuffer is configured to the same
mode (1920x1080, 4:2:0 chroma subsampling).