monitor: Compile with -mno-dsp
By default, GCC uses the MIPS DSP instruction set extension, which
doesn't seem to work under some conditions, notably when running as boot1
(DSP probably has to be enabled somehow).

The symptom was that the program crashed when executing a switch
statement, because the jump table lookup generated by the compiler used
the LWX instruction.